requirements.txt
beautifulsoup4
requests
pandas
sentence-transformers
faiss-cpu
torch
transformers
lxml
ctransformers
numpy